All base vector, matrix, and quaternion math complete

+++ src/datatypes/int3.cpp
@@ -1,4 +1,7 @@
+#include <datatype/int3.h>
+
 #include <maths/math.h>
+
 #include <utility/managed_memory_block.h>
 
 using namespace eloo;
@@ -6,22 +9,22 @@ using namespace eloo;
 namespace {
     constexpr int MEMORY_BLOCK_INITIAL_SIZE = 50;
     constexpr float MEMORY_BLOCK_EXPANSION = 0.7f;
-    using memblock_t = managed_memory_block<float, MEMORY_BLOCK_INITIAL_SIZE, MEMORY_BLOCK_EXPANSION>;
+    using memblock_t = managed_memory_block<int, MEMORY_BLOCK_INITIAL_SIZE, MEMORY_BLOCK_EXPANSION>;
     static memblock_t gMemoryBlockX, gMemoryBlockY, gMemoryBlockZ;
 }
 
-float3::id_t float3::create(float x, float y, float z, bool useIDPool) {
+int3::id_t int3::create(int x, int y, int z, bool useIDPool) {
     id_t id = gMemoryBlockX.push(x, useIDPool);
     ELOO_ASSERT_FATAL(id == gMemoryBlockY.push(y, useIDPool), "ID mismatch between X and Y memory blocks");
     ELOO_ASSERT_FATAL(id == gMemoryBlockZ.push(z, useIDPool), "ID mismatch between X and Z memory blocks");
     return id;
 }
 
-float3::id_t float3::create(const values& vals, bool useIDPool) {
+int3::id_t int3::create(const values& vals, bool useIDPool) {
     return create(vals.x(), vals.y(), vals.z(), useIDPool);
 }
 
-bool float3::try_release(id_t id) {
+bool int3::try_release(id_t id) {
     if (!is_valid(id)) {
         return false;
     }
@@ -31,7 +34,7 @@ bool float3::try_release(id_t id) {
     return true;
 }
 
-bool float3::try_get_values(id_t id, float3::values& vals) {
+bool int3::try_get_values(id_t id, int3::values& vals) {
     if (is_valid(id)) {
         vals = {
             gMemoryBlockX.get(id),
@@ -43,48 +46,48 @@ bool float3::try_get_values(id_t id, float3::values& vals) {
     return false;
 }
 
-float& float3::x(id_t id) { return gMemoryBlockX.get(id); }
-float& float3::y(id_t id) { return gMemoryBlockY.get(id); }
-float& float3::z(id_t id) { return gMemoryBlockZ.get(id); }
+int& int3::x(id_t id) { return gMemoryBlockX.get(id); }
+int& int3::y(id_t id) { return gMemoryBlockY.get(id); }
+int& int3::z(id_t id) { return gMemoryBlockZ.get(id); }
 
-const float& float3::const_x(id_t id) { return gMemoryBlockX.get(id); }
-const float& float3::const_y(id_t id) { return gMemoryBlockY.get(id); }
-const float& float3::const_z(id_t id) { return gMemoryBlockZ.get(id); }
+const int& int3::const_x(id_t id) { return gMemoryBlockX.get(id); }
+const int& int3::const_y(id_t id) { return gMemoryBlockY.get(id); }
+const int& int3::const_z(id_t id) { return gMemoryBlockZ.get(id); }
 
 
 /////////////////////////////////////////////////////////////////
-// float3::values
+// int3::values
 
-bool float3::operator != (const float3::values& lhs, const float3::values& rhs) {
+bool int3::operator != (const int3::values& lhs, const int3::values& rhs) {
     return
         lhs.x() != rhs.x() ||
         lhs.y() != rhs.y() ||
         lhs.z() != rhs.z();
 }
-bool float3::operator != (const float3::values& lhs, float rhs) {
+bool int3::operator != (const int3::values& lhs, int rhs) {
     return
         lhs.x() != rhs ||
         lhs.y() != rhs ||
         lhs.z() != rhs;
 }
 
-bool float3::operator == (const float3::values& lhs, const float3::values& rhs) {
+bool int3::operator == (const int3::values& lhs, const int3::values& rhs) {
     return
         lhs.x() == rhs.x() &&
         lhs.y() == rhs.y() &&
         lhs.z() == rhs.z();
 }
-bool float3::operator == (const float3::values& lhs, float rhs) {
+bool int3::operator == (const int3::values& lhs, int rhs) {
     return
         lhs.x() == rhs &&
         lhs.y() == rhs &&
         lhs.z() == rhs;
 }
 
-float3::values float3::operator + (const float3::values& lhs) {
+int3::values int3::operator + (const int3::values& lhs) {
     return lhs;
 }
-float3::values float3::operator - (const float3::values& lhs) {
+int3::values int3::operator - (const int3::values& lhs) {
     return {
         -lhs.x(),
         -lhs.y(),
@@ -92,14 +95,14 @@ float3::values float3::operator - (const float3::values& lhs) {
     };
 }
 
-float3::values float3::operator / (const float3::values& lhs, const float3::values& rhs) {
+int3::values int3::operator / (const int3::values& lhs, const int3::values& rhs) {
     return {
         lhs.x() / rhs.x(),
         lhs.y() / rhs.y(),
         lhs.z() / rhs.z()
     };
 }
-float3::values float3::operator / (const float3::values& lhs, float rhs) {
+int3::values int3::operator / (const int3::values& lhs, int rhs) {
     return {
         lhs.x() / rhs,
         lhs.y() / rhs,
@@ -107,14 +110,14 @@ float3::values float3::operator / (const float3::values& lhs, float rhs) {
     };
 }
 
-float3::values float3::operator * (const float3::values& lhs, const float3::values& rhs) {
+int3::values int3::operator * (const int3::values& lhs, const int3::values& rhs) {
     return {
         lhs.x() * rhs.x(),
         lhs.y() * rhs.y(),
         lhs.z() * rhs.z()
     };
 }
-float3::values float3::operator * (const float3::values& lhs, float rhs) {
+int3::values int3::operator * (const int3::values& lhs, int rhs) {
     return {
         lhs.x() * rhs,
         lhs.y() * rhs,
@@ -122,14 +125,14 @@ float3::values float3::operator * (const float3::values& lhs, float rhs) {
     };
 }
 
-float3::values float3::operator + (const float3::values& lhs, const float3::values& rhs) {
+int3::values int3::operator + (const int3::values& lhs, const int3::values& rhs) {
     return {
         lhs.x() + rhs.x(),
         lhs.y() + rhs.y(),
         lhs.z() + rhs.z()
     };
 }
-float3::values float3::operator + (const float3::values& lhs, float rhs) {
+int3::values int3::operator + (const int3::values& lhs, int rhs) {
     return {
         lhs.x() + rhs,
         lhs.y() + rhs,
@@ -137,14 +140,14 @@ float3::values float3::operator + (const float3::values& lhs, float rhs) {
     };
 }
 
-float3::values float3::operator - (const float3::values& lhs, const float3::values& rhs) {
+int3::values int3::operator - (const int3::values& lhs, const int3::values& rhs) {
     return {
         lhs.x() - rhs.x(),
         lhs.y() - rhs.y(),
         lhs.z() - rhs.z()
     };
 }
-float3::values float3::operator - (const float3::values& lhs, float rhs) {
+int3::values int3::operator - (const int3::values& lhs, int rhs) {
     return {
         lhs.x() - rhs,
         lhs.y() - rhs,
@@ -154,54 +157,54 @@ float3::values float3::operator - (const float3::values& lhs, float rhs) {
 
 
 ///////////////////////////////////////////////////////
-// Assignment and manipulation float3::operators
+// Assignment and manipulation int3::operators
 
-float3::values& float3::values::operator = (const values& other) {
+int3::values& int3::values::operator = (const values& other) {
     *this = other;
     return *this;
 }
-float3::values& float3::values::operator = (float val) {
+int3::values& int3::values::operator = (int val) {
     *this = val;
     return *this;
 }
 
-float3::values& float3::values::operator - () {
+int3::values& int3::values::operator - () {
     *this = -*this;
     return *this;
 }
 
-float3::values& float3::values::operator /= (const values& other) {
+int3::values& int3::values::operator /= (const values& other) {
     *this = *this / other;
     return *this;
 }
-float3::values& float3::values::operator /= (float val) {
+int3::values& int3::values::operator /= (int val) {
     *this = *this / val;
     return *this;
 }
 
-float3::values& float3::values::operator *= (const values& other) {
+int3::values& int3::values::operator *= (const values& other) {
     *this = *this * other;
     return *this;
 }
-float3::values& float3::values::operator *= (float val) {
+int3::values& int3::values::operator *= (int val) {
     *this = *this * val;
     return *this;
 }
 
-float3::values& float3::values::operator += (const values& other) {
+int3::values& int3::values::operator += (const values& other) {
     *this = *this + other;
     return *this;
 }
-float3::values& float3::values::operator += (float val) {
+int3::values& int3::values::operator += (int val) {
     *this = *this + val;
     return *this;
 }
 
-float3::values& float3::values::operator -= (const values& other) {
+int3::values& int3::values::operator -= (const values& other) {
     *this = *this - other;
     return *this;
 }
-float3::values& float3::values::operator -= (float val) {
+int3::values& int3::values::operator -= (int val) {
     *this = *this - val;
     return *this;
 }
